requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.9.1.post1
asgiref==3.2.10
asttokens==2.4.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.3
black==22.12.0
boto3==1.34.41
botocore==1.34.41
branchio==0.1.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.1
django-apscheduler==0.6.0
django-braces==1.15.0
django-cors-headers==3.10.1
django-filter==2.3.0
django-model-utils==4.2.0
django-oauth-toolkit==1.6.0
django-phone-verify==2.0.1
django-phonenumber-field==6.3.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-social-oauth2==2.1.3
drf-yasg==1.20.0
ecdsa==0.18.0
exceptiongroup==1.2.1
executing==2.0.1
f90nml==1.4.3
fcm-django==2.2.1
firebase==4.0.1
firebase-admin==6.5.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.33.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.17.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
inapppy==2.5.2
inflection==0.5.1
ipython==8.24.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==0.4.3
nexmo==2.5.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
pandas==2.1.1
parso==0.8.4
pathspec==0.10.3
pexpect==4.9.0
phonenumbers==8.13.3
platformdirs==2.6.0
prompt-toolkit==3.0.43
proto-plus==1.22.1
protobuf==4.21.12
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.2
python-http-client==3.3.4
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.10.1
sendgrid==6.9.3
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.3
starkbank-ecdsa==2.2.0
swapper==1.3.0
tomli==2.0.1
traitlets==5.14.3
twilio==7.16.0
typing_extensions==4.11.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
uWSGI==2.0.21
wcwidth==0.2.13
wrapt==1.14.1
yarl==1.8.2
